Base-learner factory for categorical feature on a binary base-learner basis

BaselearnerCategoricalBinary can be used to estimate effects of one category of a categorical feature. The base-learner gets the data as index vector of the observations assigned to the group. For example, if the vector is (a, a, b, b, a, b), then the index vector is (1, 2, 5) for group a. This reduces memory load and fasten the fitting process.

Format

S4 object.

Usage

BaselearnerCategoricalBinary$new(data_source, list(n_obs))

arguments

data_source [data object]

data object of class CategoricalData which contains the source data.

Fields

This class doesn't contain public fields.

Methods

getData()

Get the data matrix of the target data which is used for modeling.

transformData(X)

Transform a data matrix as defined within the factory. The argument has to be a matrix with one column. In case of the categorical binary base-learner this is the index of non-zero elements concatinated with the number of observations. This helps to fully reconstruct the original feature by using less memory. This also speed up computation time.

summarizeFactory()

Summarize the base-learner factory object.

Examples

# Sample data:
x = sample(c("pos","neg"), 20, TRUE)

# Create new data object:
data_source = CategoricalData$new(x, "pos")

# Create new linear base-learner:
cat_factory = BaselearnerCategoricalBinary$new(data_source, "pos")

# Get the transformed data as stored for internal use:
cat_factory$getData()

#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    0
#>  [4,]    1
#>  [5,]    1
#>  [6,]    0
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    0
#> [11,]    0
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    0
#> [17,]    1
#> [18,]    1
#> [19,]    0
#> [20,]    0

# Summarize factory:
cat_factory$summarizeFactory()

#> Categorical factory of category pos