Base-learner factory to do non-parametric B or P-spline regression

BaselearnerPSpline creates a spline base-learner factory object which can be registered within a base-learner list and then used for training.

Format

S4 object.

Usage

BaselearnerPSpline$new(data_source, list(degree, n_knots, penalty,
  differences, df, n_bins, bin_method))

arguments

data_source [data object]

data object which contains the source data.

degree [integer(1)]

degree of the spline functions to interpolate the knots.

n_knots [integer(1)]

number of inner knots. to prevent weird behavior on the edges the inner knots are expanded by \(\mathrm{degree} - 1\) additional knots.

penalty [numeric(1)]

positive numeric value to specify the penalty parameter. setting the penalty to 0 ordinary b-splines are used for the fitting.

differences [integer(1)]

the number of differences which are penalized. a higher value leads to smoother curves.

bin_root [integer(1)]

If set to a value greater than zero, binning is applied and reduces the number of used x values to n^(1/bin_root) equidistant points. If you want to use binning we suggest to set bin_root = 2.

bin_method [character(1)]

Method used for spacing knot points. Options are linear (equally spaced grid) or quantile (knot points based on quantiles).

Details

If bin_root > 0 the original feature is discretized to on an equidistant grid on \([\min(x),\max(x)]\) with \(\sqrt{n}\) points. The fitting is then done by using weights per new data point.

Fields

This class doesn't contain public fields.

Methods

getData()

Get the data matrix of the target data which is used for modeling.

summarizeFactory()

Summarize the base-learner factory object.

Examples

# Sample data:
data_mat = cbind(1:10)
y = sin(1:10)

# Create new data object:
data_source = InMemoryData$new(data_mat, "my_data_name")

# Create new linear base-learner:
spline_factory = BaselearnerPSpline$new(data_source,
  list(degree = 3, n_knots = 4, penalty = 2, differences = 2))

# Get the transformed data:
spline_factory$getData()
#>           [,1]       [,2]         [,3]       [,4]        [,5]        [,6]
#> [1,] 0.1666667 0.01463192 0.0000000000 0.00000000 0.000000000 0.000000000
#> [2,] 0.6666667 0.44375857 0.1170553269 0.00617284 0.000000000 0.000000000
#> [3,] 0.1666667 0.51303155 0.6550068587 0.37037037 0.078417924 0.001828989
#> [4,] 0.0000000 0.02857796 0.2277091907 0.57407407 0.622770919 0.296982167
#> [5,] 0.0000000 0.00000000 0.0002286237 0.04938272 0.296982167 0.622770919
#> [6,] 0.0000000 0.00000000 0.0000000000 0.00000000 0.001828989 0.078417924
#> [7,] 0.0000000 0.00000000 0.0000000000 0.00000000 0.000000000 0.000000000
#> [8,] 0.0000000 0.00000000 0.0000000000 0.00000000 0.000000000 0.000000000
#>            [,7]         [,8]       [,9]     [,10]
#> [1,] 0.00000000 0.0000000000 0.00000000 0.0000000
#> [2,] 0.00000000 0.0000000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.0000000000 0.00000000 0.0000000
#> [4,] 0.04938272 0.0002286237 0.00000000 0.0000000
#> [5,] 0.57407407 0.2277091907 0.02857796 0.0000000
#> [6,] 0.37037037 0.6550068587 0.51303155 0.1666667
#> [7,] 0.00617284 0.1170553269 0.44375857 0.6666667
#> [8,] 0.00000000 0.0000000000 0.01463192 0.1666667

# Summarize factory:
spline_factory$summarizeFactory()
#> Spline factory of degree 3
#> 	- Name of the used data: my_data_name
#> 	- Factory creates the following base-learner: spline_degree_3