Base-learner factory for polynomial regression

BaselearnerPolynomial creates a polynomial base-learner factory object which can be registered within a base-learner list and then used for training.

Format

S4 object.

Usage

BaselearnerPolynomial$new(data_source, list(degree, intercept))
BaselearnerPolynomial$new(data_source, blearner_type, list(degree, intercept))

Arguments

data_source [Data Object]

Data object which contains the source data.

degree [integer(1)]

This argument is used for transforming the source data. Each element is taken to the power of the degree argument.

intercept [logical(1)]

Indicating whether an intercept should be added or not. Default is set to TRUE.

Details

The polynomial base-learner factory takes any matrix which the user wants to pass the number of columns indicates how much parameter are estimated.

Fields

This class doesn't contain public fields.

Methods

getData()

Get the data matrix of the target data which is used for modeling.

summarizeFactory()

Summarize the base-learner factory object.

Examples

# Sample data:
data_mat = cbind(1:10)

# Create new data object:
data_source = InMemoryData$new(data_mat, "my_data_name")

# Create new linear base-learner factory:
lin_factory = BaselearnerPolynomial$new(data_source,
  list(degree = 2, intercept = FALSE))
lin_factory_int = BaselearnerPolynomial$new(data_source,
  list(degree = 2, intercept = TRUE))

# Get the transformed data:
lin_factory$getData()
#>       [,1] [,2]
#>  [1,]    1    1
#>  [2,]    2    4
#>  [3,]    3    9
#>  [4,]    4   16
#>  [5,]    5   25
#>  [6,]    6   36
#>  [7,]    7   49
#>  [8,]    8   64
#>  [9,]    9   81
#> [10,]   10  100
lin_factory_int$getData()
#>       [,1] [,2] [,3]
#>  [1,]    1    1    1
#>  [2,]    1    2    4
#>  [3,]    1    3    9
#>  [4,]    1    4   16
#>  [5,]    1    5   25
#>  [6,]    1    6   36
#>  [7,]    1    7   49
#>  [8,]    1    8   64
#>  [9,]    1    9   81
#> [10,]    1   10  100

# Summarize factory:
lin_factory$summarizeFactory()
#> Quadratic base-learner factory:
#> 	- Name of the used data: my_data_name
#> 	- Factory creates the following base-learner: polynomial_degree_2