Compboost API

Compboost wraps the S4 class system exposed by Rcpp to make defining objects, adding objects, the training, calculating predictions, and plotting much easier. As already mentioned, the Compboost R6 class is just a wrapper and compatible with the most S4 classes.

Format

R6Class object.

Usage

# Constructor

cboost = Compboost$new(data, target, optimizer = OptimizerCoordinateDescent$new(), loss,
  learning_rate = 0.05, oob_fraction = NULL)

# Member functions

cboost$addLogger(logger, use_as_stopper = FALSE, logger_id, ...)

cboost$addBaselearner(feature, id, bl_factory, data_source = InMemoryData, ...)

cboost$addTensor(features, id, data_source = InMemoryData, ...)

cboost$addComponents(feature, id, data_source = InMemoryData, ...)

cboost$train(iteration = 100, trace = -1)

cboost$getCurrentIteration()

cboost$prepareData(newdata)

cboost$prepareResponse(response)

cboost$predict(newdata = NULL, as_response = FALSE)

cboost$getInbagRisk()

cboost$getSelectedBaselearner()

cboost$getEstimatedCoef()

cboost$getBaselearnerNames()

cboost$getLoggerData()

cboost$calculateFeatureImportance(num_feats = NULL)

Arguments

For Compboost$new():

data

[data.frame]
A data frame containing the data (features as well as target).

target

[character(1) or S4 Response]
Character value containing the target variable or Response object. Note that the loss has to match the data type of the target.

optimizer

[S4 Optimizer]
An initialized S4 Optimizer object exposed by Rcpp (e.g. OptimizerCoordinateDescent$new()) to specify how features are selected in each iteration.

loss

[S4 Loss]
Initialized S4 Loss object exposed by Rcpp which is used to calculate the risk and pseudo residuals (e.g. LossQuadratic$new()).

learning_rage

[numeric(1)]
Learning rate to shrink the new parameters in each iteration.

oob_fraction

[numeric(1)]
Fraction of how much data are used to calculate the out of bag risk.

stop_args

[list(2)]
List containing two elements `patience` and `eps_for_break` which can be set to use early stopping on the left out data from setting `oob_fraction`.

For cboost$addLogger():

logger

[S4 Logger]
Uninitialized S4 Logger class object that is registered in the model. See the details for possible choices.

use_as_stopper

[logical(1)]
Logical value indicating whether the new logger should also be used as stopper (early stopping). Default value is FALSE.

logger_id

[character(1)]
Id of the new logger. This is necessary to be able to register multiple logger.

...
Further arguments passed to the constructor of the S4 Logger class specified in logger. For possible arguments see details or the help pages (e.g. ?LoggerIteration).

For cboost$addBaselearner():

feature

[character()]
Vector of column names that are used as input data matrix for a single base-learner. Note that not every base-learner supports the use of multiple features (e.g. the spline base-learner does not).

id

[character(1)]
Id of the base-learners. This is necessary since it is possible to define multiple learners using equal features.

bl_factory

[S4 Factory]
Uninitialized base-learner factory given as S4 Factory class. See the details for possible choices.

data_source

[S4 Data]
Data source object. Just in memory data objects are supported at the moment.

...
Further arguments passed to the constructor of the S4 Factory class specified in bl_factory. For possible arguments see the help pages (e.g. ?BaselearnerPSplineFactory) of the S4 classes.

For cboost$train():

iteration

[integer(1)]
Number of iterations that are trained. If the model is already trained it sets to the given number by going back to already trained base-learners or it trains new ones. Note: This function defines an iteration logger with the id _iterations which is used as stopper for the new training.

trace

[integer(1)]
Integer indicating after how many iterations a trace should be printed. Specifying trace = 10, then every 10th iteration is printed. If you do not want to print the trace set trace = 0. Default is -1 which means that in total 40 iterations are printed.

For cboost$predict():

newdata

[data.frame()]
Data to predict on. If newdata equals NULL predictions on the training data are returned.

Details

Loss
Available choices for the loss are:

LossQuadratic (Regression)

LossAbsolute (Regression)

LossQuantile (Regression)

quantile

[numeric(1)]
Quantile that is boosted.

LossHuber (Regression)

delta

[numeric(1)]
Defining the interval [-d,d] around 0 for quadratic approximation.

LossBinomial (Binary Classification)

LossCustom (Custom)

(For each loss take also a look at the help pages (e.g. ?LossBinomial))

Logger
Available choices for the logger are:

LoggerIteration: Logs the current iteration. Additional arguments:

max_iterations [integer(1)]

Maximal number of iterations.

LoggerTime: Logs the elapsed time. Additional arguments:

max_time [integer(1)]

Maximal time for the computation.

time_unit [character(1)]

Character to specify the time unit. Possible choices are minutes, seconds, or microseconds.

LoggerInbagRisk:

used_loss [S4 Loss]

Loss as initialized S4 Loss which is used to calculate the empirical risk. See the details for possible choices.

eps_for_break [numeric(1)]

This argument is used if the logger is also used as stopper. If the relative improvement of the logged inbag risk falls below this boundary, then the stopper breaks the algorithm.

patience [integer(1)]

Specifying, how many iteration should fall consecutively below eps_for_break before we stop.

LoggerOobRisk:

used_loss [S4 Loss]

Loss as initialized S4 Loss which is used to calculate the empirical risk. See the details for possible choices.

eps_for_break [numeric(1)]

This argument is used if the logger is also used as stopper. If the relative improvement of the logged inbag risk falls above this boundary the stopper breaks the algorithm.

oob_data [list]

A list which contains data source objects which corresponds to the source data of each registered factory. The source data objects should contain the out of bag data. This data is then used to calculate the new predictions in each iteration.

oob_response [vector]

Vector which contains the response for the out of bag data given within oob_data.

patience [integer(1)]

Specifying, how many iteration should fall consecutively below eps_for_break before we stop.

Note:

Even if you do not use the logger as stopper you have to define the arguments such as max_time.

Fields

data [data.frame]

Data used for training the algorithm.

data_oob [data.frame]

Data used for out of bag tracking.

oob_fraction [numeric(1)]

Fraction of how much data are used to track the out of bag risk.

response [vector]

Response object that is created or passed in target for training the model.

response_oob [vector]

Response object that is created by specifying the oob_fraction to evaluate each iteration.

target [character(1)]

Name of the target variable.

id [character(1)]

Name of the given dataset.

optimizer [S4 Optimizer]

Optimizer used within the fitting process.

loss [S4 Loss]

Loss used to calculate pseudo residuals and empirical risk.

learning_rate [numeric(1)]

Learning rate used to shrink the estimated parameter in each iteration.

model [S4 Compboost_internal]

S4 Compboost_internal class object from which the main operations (such as train) are called.

bl_factory_list [S4 FactoryList]

List of all registered factories represented as S4 FactoryList class.

positive_category [character(1)]

Character containing the name of the positive class in the case of (binary) classification.

stop_if_all_stoppers_fulfilled [logical(1)]

Logical indicating whether all stopper should be used simultaneously or if it is sufficient to just use the first stopper to stop the algorithm.

Methods

addLogger

method to add a logger to the algorithm (Note: This is just possible before the training).

addBaselearner

method to add a new base-learner to the algorithm (Note: This is just possible before the training).

getCurrentIteration

method to get the current iteration on which the algorithm is set.

train

method to train the algorithm.

predict

method to predict on a trained object.

getSelectedBaselearner

method to get a character vector of selected base-learner.

getEstimatedCoef

method to get a list of estimated coefficient of each selected base-learner.

getBaselearnerNames

method to get the names of the registered factories.

prepareData

method to prepare data to track the out of bag risk of an arbitrary loss/performance function.

getLoggerData

method to the the logged data from all registered logger.

calculateFeatureImportance

method to calculate feature importance.

Examples

cboost = Compboost$new(mtcars, "mpg", loss = LossQuadratic$new(), oob_fraction = 0.3)
cboost$addBaselearner("hp", "spline", BaselearnerPSpline, degree = 3,
  n_knots = 10, penalty = 2, differences = 2)
cboost$addBaselearner("wt", "spline", BaselearnerPSpline)
cboost$train(1000)
#>    1/1000   risk = 18  oob_risk = 13   
#>   25/1000   risk = 3.5  oob_risk = 2   
#>   50/1000   risk = 1.8  oob_risk = 1.5   
#>   75/1000   risk = 1.5  oob_risk = 1.7   
#>  100/1000   risk = 1.4  oob_risk = 1.9   
#>  125/1000   risk = 1.3  oob_risk = 2   
#>  150/1000   risk = 1.3  oob_risk = 2.1   
#>  175/1000   risk = 1.2  oob_risk = 2.2   
#>  200/1000   risk = 1.2  oob_risk = 2.3   
#>  225/1000   risk = 1.1  oob_risk = 2.4   
#>  250/1000   risk = 1.1  oob_risk = 2.5   
#>  275/1000   risk = 1  oob_risk = 2.6   
#>  300/1000   risk = 1  oob_risk = 2.7   
#>  325/1000   risk = 0.97  oob_risk = 2.8   
#>  350/1000   risk = 0.95  oob_risk = 3   
#>  375/1000   risk = 0.92  oob_risk = 3.1   
#>  400/1000   risk = 0.9  oob_risk = 3.2   
#>  425/1000   risk = 0.88  oob_risk = 3.3   
#>  450/1000   risk = 0.86  oob_risk = 3.4   
#>  475/1000   risk = 0.85  oob_risk = 3.5   
#>  500/1000   risk = 0.83  oob_risk = 3.6   
#>  525/1000   risk = 0.82  oob_risk = 3.7   
#>  550/1000   risk = 0.8  oob_risk = 3.8   
#>  575/1000   risk = 0.79  oob_risk = 3.9   
#>  600/1000   risk = 0.78  oob_risk = 4   
#>  625/1000   risk = 0.77  oob_risk = 4.1   
#>  650/1000   risk = 0.76  oob_risk = 4.2   
#>  675/1000   risk = 0.75  oob_risk = 4.3   
#>  700/1000   risk = 0.74  oob_risk = 4.3   
#>  725/1000   risk = 0.73  oob_risk = 4.4   
#>  750/1000   risk = 0.72  oob_risk = 4.5   
#>  775/1000   risk = 0.72  oob_risk = 4.6   
#>  800/1000   risk = 0.71  oob_risk = 4.7   
#>  825/1000   risk = 0.7  oob_risk = 4.7   
#>  850/1000   risk = 0.7  oob_risk = 4.8   
#>  875/1000   risk = 0.69  oob_risk = 4.9   
#>  900/1000   risk = 0.68  oob_risk = 4.9   
#>  925/1000   risk = 0.68  oob_risk = 5   
#>  950/1000   risk = 0.67  oob_risk = 5.1   
#>  975/1000   risk = 0.67  oob_risk = 5.1   
#> 1000/1000   risk = 0.66  oob_risk = 5.2   
#> 
#> 
#> Train 1000 iterations in 0 Seconds.
#> Final risk based on the train set: 0.66
#> 

table(cboost$getSelectedBaselearner())
#> 
#> hp_spline wt_spline 
#>       280       720