0-1 Loss for binary classification derived of the binomial distribution

This loss can be used for binary classification. The coding we have chosen here acts on \(y \in \{-1, 1\}\).

Format

S4 object.

Details

Loss Function: $$ L(y, f(x)) = \log(1 + \mathrm{exp}(-2yf(x))) $$ Gradient: $$ \frac{\delta}{\delta f(x)}\ L(y, f(x)) = - \frac{y}{1 + \mathrm{exp}(2yf)} $$ Initialization: $$ \hat{f}^{[0]}(x) = \frac{1}{2}\mathrm{log}(p / (1 - p)) $$ with $$ p = \frac{1}{n}\sum\limits_{i=1}^n\mathrm{1}_{\{y^{(i)} = 1\}} $$

Usage

LossBinomial$new()
LossBinomial$new(offset)

Arguments

offset [numeric(1)]

Numerical value which can be used to set a custom offset. If so, this value is returned instead of the loss optimal initialization.

Examples

# Create new loss object:
bin_loss = LossBinomial$new()
bin_loss

#> 
#> LossBinomial Loss:
#> 
#>   Loss function: L(y,x) = log(1 + exp(-2yf(x))
#> 
#>   
#>