Coordinate Descent with Cosine Annealing

This class defines a new object which is used to conduct Coordinate Descent with a cosine annealing learning rate strategy.

Format

S4 object.

Usage

OptimizerCosineAnnealing$new()
OptimizerCosineAnnealing$new(ncores)
OptimizerCosineAnnealing$new(nu_min, nu_max, cycles, anneal_iter_max, cycles)
OptimizerCosineAnnealing$new(nu_min, nu_max, cycles, anneal_iter_max, cycles, ncores)

Arguments

nu_min [numeric(1)]

Minimal learning rate.

nu_max [numeric(1)]

Maximal learning rate.

cycles [integer(1)]

Number of annealings form nu_max to nu_min between 1 and anneal_iter_max.

anneal_iter_max [integer(1)]

Maximal number of iters for which annealing is applied. If the iteration is bigger than anneal_iter_max, then nu_min is used as fixed learning rate.

ncores [integer(1)]

Number of cores used to fit the algorithm. Note that number of used cores should be smaller or equal the number of base learner.

Examples

# Define optimizer:
optimizer = OptimizerCosineAnnealing$new()