Wrapper to boost linear models for each feature.

This wrapper function automatically initializes the model by adding all numerical features as linear base-learner. Categorical features are dummy encoded and inserted using another linear base-learners without intercept. The function boostLinear does also train the model.

Usage

boostLinear(
  data,
  target,
  optimizer = NULL,
  loss = NULL,
  learning_rate = 0.05,
  iterations = 100,
  trace = -1,
  intercept = TRUE,
  data_source = InMemoryData,
  oob_fraction = NULL
)

Arguments

data

data.frame
A data frame containing the data.

target

character(1) or Response class
Character value containing the target variable or response object. Note that the loss must match the data type of the target.

optimizer

S4 Optimizer
An initialized S4 Optimizer object exposed by Rcpp (e.g. OptimizerCoordinateDescent$new()) to select features at each iteration.

loss

S4 Loss
Initialized S4 Loss object exposed by Rcpp that is used to calculate the risk and pseudo residuals (e.g. LossQuadratic$new()).

learning_rate

numeric(1)
Learning rate to shrink the parameter in each step.

iterations

integer(1)
Number of iterations that are trained. If iterations == 0, the untrained object is returned. This can be useful if other base learners (e.g. an interaction via a tensor base learner) are added.

trace

integer(1)
Integer indicating how often a trace should be printed. Specifying trace = 10, then every 10th iteration is printed. If no trace should be printed set trace = 0. Default is -1 which means that in total 40 iterations are printed.

intercept

logical(1)
Internally used by BaselearnerPolynomial. This logical value indicates if each feature should get an intercept or not (default is TRUE).

data_source

S4 Data
Uninitialized S4 Data object which is used to store the data. At the moment just in memory training is supported.

oob_fraction

numeric(1)
Fraction of how much data are used to track the out of bag risk.

Value

A model of the Compboost class. This model is an R6 object which can be used for retraining, predicting, plotting, and anything described in ?Compboost.

Details

The returned object is an object of the Compboost class. This object can be used for further analyses (see ?Compboost for details).

Examples

mod = boostLinear(data = iris, target = "Sepal.Length", loss = LossQuadratic$new(),
  oob_fraction = 0.3)
#>   1/100   risk = 0.3  oob_risk = 0.34   
#>   2/100   risk = 0.28  oob_risk = 0.31   
#>   4/100   risk = 0.24  oob_risk = 0.27   
#>   6/100   risk = 0.21  oob_risk = 0.24   
#>   8/100   risk = 0.19  oob_risk = 0.21   
#>  10/100   risk = 0.17  oob_risk = 0.19   
#>  12/100   risk = 0.15  oob_risk = 0.17   
#>  14/100   risk = 0.14  oob_risk = 0.16   
#>  16/100   risk = 0.12  oob_risk = 0.14   
#>  18/100   risk = 0.12  oob_risk = 0.14   
#>  20/100   risk = 0.11  oob_risk = 0.13   
#>  22/100   risk = 0.1  oob_risk = 0.12   
#>  24/100   risk = 0.097  oob_risk = 0.12   
#>  26/100   risk = 0.093  oob_risk = 0.11   
#>  28/100   risk = 0.09  oob_risk = 0.11   
#>  30/100   risk = 0.087  oob_risk = 0.11   
#>  32/100   risk = 0.085  oob_risk = 0.1   
#>  34/100   risk = 0.083  oob_risk = 0.1   
#>  36/100   risk = 0.081  oob_risk = 0.099   
#>  38/100   risk = 0.08  oob_risk = 0.097   
#>  40/100   risk = 0.078  oob_risk = 0.095   
#>  42/100   risk = 0.077  oob_risk = 0.093   
#>  44/100   risk = 0.075  oob_risk = 0.092   
#>  46/100   risk = 0.074  oob_risk = 0.09   
#>  48/100   risk = 0.072  oob_risk = 0.089   
#>  50/100   risk = 0.071  oob_risk = 0.087   
#>  52/100   risk = 0.07  oob_risk = 0.086   
#>  54/100   risk = 0.069  oob_risk = 0.085   
#>  56/100   risk = 0.068  oob_risk = 0.083   
#>  58/100   risk = 0.067  oob_risk = 0.082   
#>  60/100   risk = 0.066  oob_risk = 0.081   
#>  62/100   risk = 0.065  oob_risk = 0.08   
#>  64/100   risk = 0.064  oob_risk = 0.079   
#>  66/100   risk = 0.063  oob_risk = 0.079   
#>  68/100   risk = 0.062  oob_risk = 0.078   
#>  70/100   risk = 0.062  oob_risk = 0.077   
#>  72/100   risk = 0.061  oob_risk = 0.076   
#>  74/100   risk = 0.06  oob_risk = 0.076   
#>  76/100   risk = 0.06  oob_risk = 0.075   
#>  78/100   risk = 0.059  oob_risk = 0.075   
#>  80/100   risk = 0.058  oob_risk = 0.074   
#>  82/100   risk = 0.058  oob_risk = 0.074   
#>  84/100   risk = 0.057  oob_risk = 0.073   
#>  86/100   risk = 0.057  oob_risk = 0.073   
#>  88/100   risk = 0.056  oob_risk = 0.072   
#>  90/100   risk = 0.056  oob_risk = 0.072   
#>  92/100   risk = 0.056  oob_risk = 0.072   
#>  94/100   risk = 0.055  oob_risk = 0.071   
#>  96/100   risk = 0.055  oob_risk = 0.071   
#>  98/100   risk = 0.054  oob_risk = 0.071   
#> 100/100   risk = 0.054  oob_risk = 0.071   
#> 
#> 
#> Train 100 iterations in 0 Seconds.
#> Final risk based on the train set: 0.054
#> 
mod$getBaselearnerNames()
#> [1] "Sepal.Width_linear"  "Petal.Length_linear" "Petal.Width_linear" 
#> [4] "Species_ridge"      
mod$getEstimatedCoef()
#> Depricated, use `$getCoef()` instead.
#> $Petal.Length_linear
#>            [,1]
#> [1,] -1.5978252
#> [2,]  0.4248459
#> attr(,"blclass")
#> [1] "Rcpp_BaselearnerPolynomial"
#> 
#> $Sepal.Width_linear
#>            [,1]
#> [1,] -1.1476025
#> [2,]  0.3766748
#> attr(,"blclass")
#> [1] "Rcpp_BaselearnerPolynomial"
#> 
#> $offset
#> [1] 5.833333
#> 
table(mod$getSelectedBaselearner())
#> 
#> Petal.Length_linear  Sepal.Width_linear 
#>                  66                  34 
mod$predict()
#>            [,1]
#>   [1,] 5.001052
#>   [2,] 4.845565
#>   [3,] 4.892866
#>   [4,] 5.038719
#>   [5,] 5.279175
#>   [6,] 4.963384
#>   [7,] 5.005869
#>   [8,] 4.775047
#>   [9,] 4.892866
#>  [10,] 5.048353
#>  [11,] 4.685260
#>  [12,] 5.104420
#>  [13,] 5.382544
#>  [14,] 5.109237
#>  [15,] 5.241508
#>  [16,] 5.156539
#>  [17,] 5.118871
#>  [18,] 4.868781
#>  [19,] 5.175807
#>  [20,] 4.897683
#>  [21,] 5.048353
#>  [22,] 5.043536
#>  [23,] 4.973018
#>  [24,] 4.935351
#>  [25,] 5.269541
#>  [26,] 4.958567
#>  [27,] 5.038719
#>  [28,] 4.770230
#>  [29,] 5.005869
#>  [30,] 4.958567
#>  [31,] 4.845565
#>  [32,] 5.086021
#>  [33,] 4.812714
#>  [34,] 4.925717
#>  [35,] 6.205071
#>  [36,] 6.096886
#>  [37,] 6.327708
#>  [38,] 5.393917
#>  [39,] 6.134554
#>  [40,] 6.002283
#>  [41,] 5.615974
#>  [42,] 6.177038
#>  [43,] 5.709708
#>  [44,] 6.124919
#>  [45,] 6.129737
#>  [46,] 5.846796
#>  [47,] 5.828397
#>  [48,] 5.686492
#>  [49,] 6.332525
#>  [50,] 6.111337
#>  [51,] 6.139371
#>  [52,] 6.007100
#>  [53,] 6.087252
#>  [54,] 6.342159
#>  [55,] 6.092069
#>  [56,] 5.554221
#>  [57,] 5.563855
#>  [58,] 5.761827
#>  [59,] 6.271642
#>  [60,] 6.252373
#>  [61,] 5.823580
#>  [62,] 5.959798
#>  [63,] 5.936582
#>  [64,] 6.172221
#>  [65,] 5.356249
#>  [66,] 5.889280
#>  [67,] 5.964615
#>  [68,] 6.007100
#>  [69,] 5.304130
#>  [70,] 5.884463
#>  [71,] 6.271642
#>  [72,] 6.724521
#>  [73,] 5.941399
#>  [74,] 6.856792
#>  [75,] 6.493699
#>  [76,] 7.035495
#>  [77,] 6.356611
#>  [78,] 6.554582
#>  [79,] 6.153822
#>  [80,] 6.554582
#>  [81,] 7.365737
#>  [82,] 6.998697
#>  [83,] 6.714887
#>  [84,] 6.224340
#>  [85,] 6.186672
#>  [86,] 6.752554
#>  [87,] 6.842340
#>  [88,] 6.181855
#>  [89,] 6.299675
#>  [90,] 6.521732
#>  [91,] 6.682036
#>  [92,] 6.521732
#>  [93,] 6.309309
#>  [94,] 6.446397
#>  [95,] 6.809490
#>  [96,] 6.592250
#>  [97,] 6.257190
#>  [98,] 6.549765
#>  [99,] 6.634734
#> [100,] 6.271642
#> [101,] 6.752554
#> [102,] 6.427129
#> [103,] 6.153822
#> [104,] 6.427129
#> [105,] 6.384644