Get C++ example script to define a custom cpp logger

This function can be used to print the trace of the parameters of a trained compboost object.

getCustomCppExample(example = "blearner", silent = FALSE)

Arguments

example
[character(1)]
Character value indicating if an example for the base-learner or for the loss should be returned. The values has to be one of blearner or loss.

silent
[logical(1)]
Logical value indicating if the example code should be printed to the screen or not.

Value

This function returns a character vector that can be compiled using Rcpp::sourceCpp(code = getCustomCppExample()) to define a new custom cpp logger.