Classification component-wise boosting learner

A LearnerClassif for a component-wise boosting model implemented in Compboost in package compboost.

Initial parameter values

Parameter xval is initialized to 0 in order to save some computation time.

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifCompboost

Methods

Public methods

LearnerClassifCompboost$new()

LearnerClassifCompboost$importance()

LearnerClassifCompboost$selected_features()

LearnerClassifCompboost$save_json()

LearnerClassifCompboost$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifCompboost$new()

Method importance()

The importance scores are extracted from the model slot $calculateFEatureImportance().

Usage

LearnerClassifCompboost$importance()

Returns

Named numeric().

Method selected_features()

Selected features are extracted from the model slot $getSelectedBaselearner().

Usage

LearnerClassifCompboost$selected_features()

Returns

character().

Method save_json()

Save the model to a JSON file.

Usage

LearnerClassifCompboost$save_json(file)

Arguments

file

(character(1)
Name/path to the file.

Returns

character(1)

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifCompboost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.