Coordinate descent with cosine annealing

Same as OptimizerCoordinateDescent but with a cosine annealing scheduler to adjust the learning rate during the fitting process.

Format

S4 object.

Arguments

ncores

(integer(1))
Number of cores over which the base learner fitting is distributed.

nu_min

(numeric(1))
Minimal learning rate.

nu_max

(numeric(1))
Maximal learning rate.

cycles

(integer(1))
Number of annealing cycles form nu_max to nu_min between 1 and anneal_anneal_iter_max. anneal_iter_max (integer(1))\cr Maximal number of iterations for which the annealing is conducted. nu_minis used as fixed learning rate afteranneal_iter_max`.

Usage

OptimizerCosineAnnealing$new()
OptimizerCosineAnnealing$new(ncores)
OptimizerCosineAnnealing$new(nu_min, nu_max, cycles, anneal_iter_max, cycles)
OptimizerCosineAnnealing$new(nu_min, nu_max, cycles, anneal_iter_max, cycles, ncores)

Fields

This class doesn't contain public fields.

Methods

$getOptimizerType(): () -> character(1)

$getStepSize(): () -> numeric()

Examples

# Define optimizer:
optimizer = OptimizerCosineAnnealing$new()