Wrapper to boost general additive models for each feature.

This wrapper function automatically initializes the model by adding all numerical features as spline base-learner. Categorical features are dummy encoded and inserted using another linear base-learners without intercept. The function boostSplines does also train the model.

Usage

boostSplines(
  data,
  target,
  optimizer = NULL,
  loss = NULL,
  learning_rate = 0.05,
  iterations = 100,
  trace = -1,
  degree = 3,
  n_knots = 20,
  penalty = 2,
  df = 0,
  differences = 2,
  data_source = InMemoryData,
  oob_fraction = NULL,
  bin_root = 0,
  cache_type = "inverse",
  stop_args = list(),
  df_cat = 1,
  stop_time = "microseconds",
  additional_risk_logs = list()
)

Arguments

data

data.frame
A data frame containing the data.

target

character(1) or Response class
Character value containing the target variable or Response object. Note that the loss must match the data type of the target.

optimizer

S4 Optimizer
An initialized S4 Optimizer object exposed by Rcpp (e.g. OptimizerCoordinateDescent$new()) to select features at each iteration.

loss

S4 Loss
Initialized S4 Loss object exposed by Rcpp that is used to calculate the risk and pseudo residuals (e.g. LossQuadratic$new()).

learning_rate

numeric(1)
Learning rate to shrink the parameter in each step.

iterations

integer(1)
Number of iterations that are trained. If iterations == 0, the untrained object is returned. This can be useful if other base learners (e.g. an interaction via a tensor base learner) are added.

trace

integer(1)
Integer indicating how often a trace should be printed. Specifying trace = 10, then every 10th iteration is printed. If no trace should be printed set trace = 0. Default is -1 which means that in total 40 iterations are printed.

degree

integer(1)
Polynomial degree of the splines.

n_knots

integer(1)
Number of equidistant "inner knots". The actual number of used knots does also depend on the polynomial degree.

penalty

numeric(1)
Penalty term for p-splines. If the penalty equals 0, then ordinary b-splines are fitted. The higher the penalty, the higher the smoothness.

df

numeric(1)
Degrees of freedom of the whole spline. It is important to set the same amount of degrees of freedom to be able to compare different base-learner.

differences

integer(1)
Number of differences that are used for penalization. The higher the difference, the higher the smoothness.

data_source

S4 Data
Uninitialized S4 Data object which is used to store the data. At the moment just in memory training is supported.

oob_fraction

numeric(1)
Fraction of how much data we want to use to track the out of bag risk.

bin_root

integer(1)
If set to a value greater than zero, binning is applied and reduces the number of used x values to n^(1/bin_root) equidistant points. If you want to use binning we suggest to set bin_root = 2.

cache_type

character(1)+
String to indicate what method should be used to estimate the parameter in each iteration. Default is cache_type = "cholesky" which computes the Cholesky decomposition, caches it, and reuses the matrix over and over again. The other option is to use cache_type = "inverse" which does the same but caches the inverse.

stop_args

list(2)
List containing two elements patience and eps_for_break which can be set to use early stopping on the left out data from setting oob_fraction.

df_cat

numeric(1)
Degrees of freedom of the categorical base-learner.

stop_time

character(1)
Unit of measured time.

additional_risk_logs

list(Logger)
Additional logger passed to the Compboost object.

Value

A model of the Compboost class. This model is an R6 object which can be used for retraining, predicting, plotting, and anything described in ?Compboost.

Details

The returned object is an object of the Compboost class. This object can be used for further analyses (see ?Compboost for details).

Examples

mod = boostSplines(data = iris, target = "Sepal.Length", loss = LossQuadratic$new(),
  oob_fraction = 0.3)
#>   1/100   risk = 0.3  oob_risk = 0.35   time = 0   
#>   2/100   risk = 0.28  oob_risk = 0.32   time = 81   
#>   4/100   risk = 0.24  oob_risk = 0.27   time = 179   
#>   6/100   risk = 0.2  oob_risk = 0.23   time = 275   
#>   8/100   risk = 0.18  oob_risk = 0.2   time = 369   
#>  10/100   risk = 0.16  oob_risk = 0.18   time = 462   
#>  12/100   risk = 0.14  oob_risk = 0.15   time = 553   
#>  14/100   risk = 0.12  oob_risk = 0.14   time = 643   
#>  16/100   risk = 0.11  oob_risk = 0.12   time = 734   
#>  18/100   risk = 0.1  oob_risk = 0.11   time = 827   
#>  20/100   risk = 0.096  oob_risk = 0.1   time = 919   
#>  22/100   risk = 0.09  oob_risk = 0.093   time = 1010   
#>  24/100   risk = 0.085  oob_risk = 0.087   time = 1129   
#>  26/100   risk = 0.08  oob_risk = 0.082   time = 1225   
#>  28/100   risk = 0.077  oob_risk = 0.078   time = 1319   
#>  30/100   risk = 0.074  oob_risk = 0.074   time = 1414   
#>  32/100   risk = 0.071  oob_risk = 0.073   time = 1530   
#>  34/100   risk = 0.069  oob_risk = 0.072   time = 1625   
#>  36/100   risk = 0.067  oob_risk = 0.072   time = 1718   
#>  38/100   risk = 0.065  oob_risk = 0.071   time = 1811   
#>  40/100   risk = 0.063  oob_risk = 0.07   time = 1904   
#>  42/100   risk = 0.061  oob_risk = 0.07   time = 1997   
#>  44/100   risk = 0.06  oob_risk = 0.069   time = 2092   
#>  46/100   risk = 0.058  oob_risk = 0.069   time = 2186   
#>  48/100   risk = 0.057  oob_risk = 0.069   time = 2280   
#>  50/100   risk = 0.056  oob_risk = 0.068   time = 2375   
#>  52/100   risk = 0.055  oob_risk = 0.068   time = 2468   
#>  54/100   risk = 0.053  oob_risk = 0.068   time = 2561   
#>  56/100   risk = 0.053  oob_risk = 0.068   time = 2657   
#>  58/100   risk = 0.052  oob_risk = 0.068   time = 2753   
#>  60/100   risk = 0.051  oob_risk = 0.068   time = 2850   
#>  62/100   risk = 0.05  oob_risk = 0.067   time = 2945   
#>  64/100   risk = 0.049  oob_risk = 0.067   time = 3041   
#>  66/100   risk = 0.049  oob_risk = 0.067   time = 3138   
#>  68/100   risk = 0.048  oob_risk = 0.067   time = 3247   
#>  70/100   risk = 0.047  oob_risk = 0.067   time = 3343   
#>  72/100   risk = 0.047  oob_risk = 0.067   time = 3439   
#>  74/100   risk = 0.046  oob_risk = 0.067   time = 3535   
#>  76/100   risk = 0.046  oob_risk = 0.067   time = 3630   
#>  78/100   risk = 0.045  oob_risk = 0.067   time = 3728   
#>  80/100   risk = 0.045  oob_risk = 0.067   time = 3824   
#>  82/100   risk = 0.044  oob_risk = 0.067   time = 3920   
#>  84/100   risk = 0.044  oob_risk = 0.067   time = 4014   
#>  86/100   risk = 0.044  oob_risk = 0.067   time = 4109   
#>  88/100   risk = 0.043  oob_risk = 0.067   time = 4206   
#>  90/100   risk = 0.043  oob_risk = 0.067   time = 4302   
#>  92/100   risk = 0.043  oob_risk = 0.067   time = 4401   
#>  94/100   risk = 0.042  oob_risk = 0.067   time = 4498   
#>  96/100   risk = 0.042  oob_risk = 0.067   time = 4595   
#>  98/100   risk = 0.042  oob_risk = 0.068   time = 4693   
#> 100/100   risk = 0.041  oob_risk = 0.067   time = 4793   
#> 
#> 
#> Train 100 iterations in 0 Seconds.
#> Final risk based on the train set: 0.041
#> 
mod$getBaselearnerNames()
#> [1] "Sepal.Width_spline"  "Petal.Length_spline" "Petal.Width_spline" 
#> [4] "Species_ridge"      
mod$getEstimatedCoef()
#> Depricated, use `$getCoef()` instead.
#> $Petal.Length_spline
#>              [,1]
#>  [1,] -1.05834736
#>  [2,] -0.90332534
#>  [3,] -0.77252565
#>  [4,] -0.62949367
#>  [5,] -0.60846779
#>  [6,] -0.69563654
#>  [7,] -0.77204494
#>  [8,] -0.80950548
#>  [9,] -0.78151531
#> [10,] -0.66157161
#> [11,] -0.43346606
#> [12,] -0.25387847
#> [13,] -0.12896417
#> [14,]  0.07727353
#> [15,]  0.31035786
#> [16,]  0.43562087
#> [17,]  0.32978932
#> [18,]  0.38060685
#> [19,]  0.51964373
#> [20,]  0.79770827
#> [21,]  1.16794500
#> [22,]  1.45643790
#> [23,]  1.58684317
#> [24,]  1.69032086
#> attr(,"blclass")
#> [1] "Rcpp_BaselearnerPSpline"
#> 
#> $Petal.Width_spline
#>                [,1]
#>  [1,] -0.2936077843
#>  [2,] -0.2306673235
#>  [3,] -0.1692863643
#>  [4,] -0.0961276922
#>  [5,] -0.0317133747
#>  [6,] -0.0349620817
#>  [7,] -0.0279065833
#>  [8,]  0.0008609872
#>  [9,]  0.0259059379
#> [10,]  0.0210443422
#> [11,] -0.0152608329
#> [12,] -0.0015115470
#> [13,]  0.0923922299
#> [14,]  0.1230523331
#> [15,]  0.0210148367
#> [16,] -0.0601193785
#> [17,] -0.0142121543
#> [18,]  0.0850542217
#> [19,]  0.1850354873
#> [20,]  0.1995192802
#> [21,]  0.1587042242
#> [22,]  0.0677795969
#> [23,]  0.0006220810
#> [24,] -0.0500095739
#> attr(,"blclass")
#> [1] "Rcpp_BaselearnerPSpline"
#> 
#> $Sepal.Width_spline
#>              [,1]
#>  [1,]  0.13190502
#>  [2,] -0.02035514
#>  [3,] -0.15822390
#>  [4,] -0.23978234
#>  [5,] -0.26619135
#>  [6,] -0.20441619
#>  [7,] -0.11714285
#>  [8,] -0.01080995
#>  [9,] -0.01873584
#> [10,]  0.07739103
#> [11,]  0.18721534
#> [12,]  0.06051406
#> [13,] -0.07708973
#> [14,] -0.09885890
#> [15,] -0.03405277
#> [16,]  0.06761781
#> [17,]  0.19453418
#> [18,]  0.29940807
#> [19,]  0.37560089
#> [20,]  0.36950815
#> [21,]  0.36624304
#> [22,]  0.39430664
#> [23,]  0.43620219
#> [24,]  0.48040270
#> attr(,"blclass")
#> [1] "Rcpp_BaselearnerPSpline"
#> 
#> $offset
#> [1] 5.845714
#> 
table(mod$getSelectedBaselearner())
#> 
#> Petal.Length_spline  Petal.Width_spline  Sepal.Width_spline 
#>                  43                  23                  34 
mod$predict()
#>            [,1]
#>   [1,] 5.007290
#>   [2,] 5.028145
#>   [3,] 4.968136
#>   [4,] 5.479036
#>   [5,] 4.941578
#>   [6,] 4.923360
#>   [7,] 4.952102
#>   [8,] 5.084884
#>   [9,] 4.952332
#>  [10,] 4.952007
#>  [11,] 5.230850
#>  [12,] 5.559389
#>  [13,] 5.345340
#>  [14,] 5.330001
#>  [15,] 5.232018
#>  [16,] 4.767411
#>  [17,] 5.182742
#>  [18,] 4.945859
#>  [19,] 5.078687
#>  [20,] 4.936908
#>  [21,] 4.880936
#>  [22,] 5.148937
#>  [23,] 5.040306
#>  [24,] 5.319954
#>  [25,] 5.336151
#>  [26,] 5.140167
#>  [27,] 4.957895
#>  [28,] 4.923360
#>  [29,] 4.830885
#>  [30,] 5.109818
#>  [31,] 5.366927
#>  [32,] 5.067933
#>  [33,] 5.115073
#>  [34,] 4.957330
#>  [35,] 6.438842
#>  [36,] 6.321046
#>  [37,] 6.456524
#>  [38,] 5.641128
#>  [39,] 6.222341
#>  [40,] 6.117392
#>  [41,] 5.086638
#>  [42,] 6.215607
#>  [43,] 5.646046
#>  [44,] 6.014023
#>  [45,] 5.739305
#>  [46,] 5.548310
#>  [47,] 6.277979
#>  [48,] 5.689221
#>  [49,] 6.188981
#>  [50,] 5.425512
#>  [51,] 5.737852
#>  [52,] 5.996854
#>  [53,] 6.187527
#>  [54,] 6.218009
#>  [55,] 6.195607
#>  [56,] 5.250892
#>  [57,] 5.400588
#>  [58,] 5.367311
#>  [59,] 6.077357
#>  [60,] 6.049422
#>  [61,] 6.447161
#>  [62,] 5.945518
#>  [63,] 5.896202
#>  [64,] 5.531823
#>  [65,] 5.827321
#>  [66,] 6.325153
#>  [67,] 5.522932
#>  [68,] 5.168299
#>  [69,] 5.792228
#>  [70,] 5.922789
#>  [71,] 5.916360
#>  [72,] 5.996854
#>  [73,] 5.805273
#>  [74,] 6.798550
#>  [75,] 6.124536
#>  [76,] 6.909171
#>  [77,] 6.786644
#>  [78,] 7.615267
#>  [79,] 5.826625
#>  [80,] 6.294212
#>  [81,] 6.937235
#>  [82,] 6.457749
#>  [83,] 6.152145
#>  [84,] 6.104067
#>  [85,] 6.214544
#>  [86,] 7.831219
#>  [87,] 6.283407
#>  [88,] 6.707694
#>  [89,] 7.525069
#>  [90,] 6.644316
#>  [91,] 6.263853
#>  [92,] 6.942510
#>  [93,] 7.618592
#>  [94,] 6.512197
#>  [95,] 6.240767
#>  [96,] 6.264662
#>  [97,] 6.357292
#>  [98,] 6.432768
#>  [99,] 6.586697
#> [100,] 6.466392
#> [101,] 6.124536
#> [102,] 6.922589
#> [103,] 6.458348
#> [104,] 6.381314
#> [105,] 6.313058