Classification component-wise boosting learner

A Learner for a component-wise boosting model implemented in Compboost in package compboost.

Super classes

mlr3::Learner -> compboost::LearnerCompboost -> LearnerClassifCompboost

Methods

Public methods

LearnerClassifCompboost$new()

LearnerClassifCompboost$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

compboost::LearnerCompboost$importance()

compboost::LearnerCompboost$save_json()

compboost::LearnerCompboost$selected_features()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifCompboost$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifCompboost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

l = lrn("classif.compboost", baselearner = "components",
  df = 5, iterations = 20)
task = tsk("german_credit")
l$train(task)
l$selected_features()
#> [1] "status_ridge"